# resources/chargebee/openapi.yaml
openapi: 3.0.3
info:
  title: Fabricate Chargebee API
  version: v2
  description: >-
    Curated Chargebee API v2 contract for the Acme site. Paths, envelopes, and
    field names follow the official Product Catalog 2.0 reference. Amounts are
    minor currency units (cents). Live Chargebee authenticates with HTTP Basic;
    this contract accepts the synthetic bearer token Fabricate injects.
servers:
  - url: https://acme.chargebee.com
security:
  - bearerAuth: []
...
paths:
  /api/v2/customers:
    get:
      operationId: listCustomers
      summary: Lists customers.
      tags: [Customers]
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/IncludeDeleted'
        - $ref: '#/components/parameters/IdIs'
        - $ref: '#/components/parameters/EmailIs'
      responses:
        '200':
          description: Customer list.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CustomerList'}
        default: {$ref: '#/components/responses/Error'}
  /api/v2/customers/{customer-id}:
    get:
      operationId: retrieveCustomer
      summary: Retrieves a customer.
      tags: [Customers]
      parameters:
        - $ref: '#/components/parameters/CustomerId'
      responses:
        '200':
          description: Customer.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CustomerEnvelope'}
        default: {$ref: '#/components/responses/Error'}
  /api/v2/subscriptions:
    get:
      operationId: listSubscriptions
      summary: Lists subscriptions.
      tags: [Subscriptions]
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/IncludeDeleted'
        - $ref: '#/components/parameters/IdIs'
        - $ref: '#/components/parameters/CustomerIdIs'
        - $ref: '#/components/parameters/StatusIs'
      responses:
        '200':
          description: Subscription list.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SubscriptionList'}
        default: {$ref: '#/components/responses/Error'}
  /api/v2/subscriptions/{subscription-id}:
    get:
      operationId: retrieveSubscription
      summary: Retrieves a subscription.
      tags: [Subscriptions]
      parameters:
        - $ref: '#/components/parameters/SubscriptionId'
      responses:
        '200':
          description: Subscription and its customer.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SubscriptionEnvelope'}
        default: {$ref: '#/components/responses/Error'}
  /api/v2/subscriptions/{subscription-id}/cancel_for_items:
    post:
      operationId: cancelSubscriptionForItems
      summary: Cancels a subscription immediately or at the end of the term.
      tags: [Subscriptions]
      parameters:
        - $ref: '#/components/parameters/SubscriptionId'
      requestBody:
        required: false
        content:
          application/x-www-form-urlencoded:
            schema: {$ref: '#/components/schemas/CancelSubscriptionRequest'}
      responses:
        '200':
          description: Cancelled or non-renewing subscription.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SubscriptionEnvelope'}
        default: {$ref: '#/components/responses/Error'}
  /api/v2/invoices:
    get:
      operationId: listInvoices
      summary: Lists invoices.
      tags: [Invoices]
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/IncludeDeleted'
        - $ref: '#/components/parameters/IdIs'
        - $ref: '#/components/parameters/CustomerIdIs'
        - $ref: '#/components/parameters/SubscriptionIdIs'
        - $ref: '#/components/parameters/StatusIs'
      responses:
        '200':
          description: Invoice list.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InvoiceList'}
        default: {$ref: '#/components/responses/Error'}
  /api/v2/invoices/{invoice-id}:
    get:
      operationId: retrieveInvoice
      summary: Retrieves an invoice.
      tags: [Invoices]
      parameters:
        - $ref: '#/components/parameters/InvoiceId'
      responses:
        '200':
          description: Invoice.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InvoiceEnvelope'}
        default: {$ref: '#/components/responses/Error'}
  /api/v2/transactions:
    get:
      operationId: listTransactions
      summary: Lists transactions.
      tags: [Transactions]
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/IncludeDeleted'
        - $ref: '#/components/parameters/IdIs'
        - $ref: '#/components/parameters/CustomerIdIs'
        - $ref: '#/components/parameters/SubscriptionIdIs'
        - $ref: '#/components/parameters/StatusIs'
      responses:
        '200':
          description: Transaction list.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TransactionList'}
        default: {$ref: '#/components/responses/Error'}
  /api/v2/transactions/{transaction-id}:
    get:
      operationId: retrieveTransaction
      summary: Retrieves a transaction.
      tags: [Transactions]
      parameters:
        - $ref: '#/components/parameters/TransactionId'
      responses:
        '200':
          description: Transaction.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TransactionEnvelope'}
        default: {$ref: '#/components/responses/Error'}
  /api/v2/comments:
    get:
      operationId: listComments
      summary: Lists comments, newest first.
      tags: [Comments]
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/EntityType'
        - $ref: '#/components/parameters/EntityId'
      responses:
        '200':
          description: Comment list.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CommentList'}
        default: {$ref: '#/components/responses/Error'}
    post:
      operationId: createComment
      summary: Creates a user comment on a customer, subscription, or invoice.
      tags: [Comments]
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema: {$ref: '#/components/schemas/CreateCommentRequest'}
      responses:
        '200':
          description: Created comment.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CommentEnvelope'}
        default: {$ref: '#/components/responses/Error'}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        Synthetic bearer token. Chargebee's live API uses HTTP Basic with the
        API key as the username and an empty password. Fabricate's proxy
        overwrites Authorization with Bearer, so this stand-in is what the
        resource accepts.
  parameters:
    Limit:
      name: limit
      in: query
      description: Page size. Chargebee's default is 10 and the maximum is 100.
      schema: {type: integer, minimum: 1, maximum: 100, default: 10}
    Offset:
      name: offset
      in: query
      description: Cursor returned as next_offset. This resource uses a decimal index.
      schema: {type: string}
    IncludeDeleted:
      name: include_deleted
      in: query
      schema: {type: boolean, default: false}
    IdIs:
      name: id[is]
      in: query
      schema: {type: string}
    EmailIs:
      name: email[is]
      in: query
      schema: {type: string}
    CustomerIdIs:
      name: customer_id[is]
      in: query
      schema: {type: string}
    SubscriptionIdIs:
      name: subscription_id[is]
      in: query
      schema: {type: string}
    StatusIs:
      name: status[is]
      in: query
      schema: {type: string}
    CustomerId:
      name: customer-id
      in: path
      required: true
      schema: {type: string, minLength: 1}
    SubscriptionId:
      name: subscription-id
      in: path
      required: true
      schema: {type: string, minLength: 1}
    InvoiceId:
      name: invoice-id
      in: path
      required: true
      schema: {type: string, minLength: 1}
    TransactionId:
      name: transaction-id
      in: path
      required: true
      schema: {type: string, minLength: 1}
    EntityType:
      name: entity_type
      in: query
      schema:
        type: string
        enum: [customer, subscription, invoice]
    EntityId:
      name: entity_id
      in: query
      schema: {type: string}
  responses:
    Error:
      description: Chargebee-style error.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
  schemas:
...
    CustomerEnvelope:
      type: object
      additionalProperties: false
      required: [customer]
      properties:
        customer: {$ref: '#/components/schemas/Customer'}
...
    CustomerList:
      type: object
      additionalProperties: false
      required: [list]
      properties:
        list:
          type: array
          items: {$ref: '#/components/schemas/CustomerListEntry'}
        next_offset: {type: string}
...
    SubscriptionEnvelope:
      type: object
      additionalProperties: false
      required: [subscription, customer]
      properties:
        subscription: {$ref: '#/components/schemas/Subscription'}
        customer: {$ref: '#/components/schemas/Customer'}
...
    SubscriptionList:
      type: object
      additionalProperties: false
      required: [list]
      properties:
        list:
          type: array
          items: {$ref: '#/components/schemas/SubscriptionListEntry'}
        next_offset: {type: string}
    CancelSubscriptionRequest:
      type: object
      additionalProperties: false
      properties:
        cancel_option:
          type: string
          enum: [immediately, end_of_term, specific_date]
        end_of_term: {type: boolean}
        cancel_at: {type: integer, format: int64}
        cancel_reason_code: {type: string, maxLength: 100}
...
    InvoiceEnvelope:
      type: object
      additionalProperties: false
      required: [invoice]
      properties:
        invoice: {$ref: '#/components/schemas/Invoice'}
...
    InvoiceList:
      type: object
      additionalProperties: false
      required: [list]
      properties:
        list:
          type: array
          items: {$ref: '#/components/schemas/InvoiceListEntry'}
        next_offset: {type: string}
...
    TransactionEnvelope:
      type: object
      additionalProperties: false
      required: [transaction]
      properties:
        transaction: {$ref: '#/components/schemas/Transaction'}
...
    TransactionList:
      type: object
      additionalProperties: false
      required: [list]
      properties:
        list:
          type: array
          items: {$ref: '#/components/schemas/TransactionListEntry'}
        next_offset: {type: string}
...
    CommentEnvelope:
      type: object
      additionalProperties: false
      required: [comment]
      properties:
        comment: {$ref: '#/components/schemas/Comment'}
...
    CommentList:
      type: object
      additionalProperties: false
      required: [list]
      properties:
        list:
          type: array
          items: {$ref: '#/components/schemas/CommentListEntry'}
        next_offset: {type: string}
    CreateCommentRequest:
      type: object
      additionalProperties: false
      required: [entity_type, entity_id, notes]
      properties:
        entity_type:
          type: string
          enum: [customer, subscription, invoice]
        entity_id: {type: string, minLength: 1, maxLength: 100}
        notes: {type: string, minLength: 1, maxLength: 1000}
        added_by: {type: string, maxLength: 100}